# Wavelet/docs/swagger.yaml
basePath: /
definitions:
...
  model.AcmeAccount:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      updated_at:
        type: string
      url:
        type: string
    type: object
...
  model.DNSAccount:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
...
  model.ManagedDomain:
    properties:
      cert_id:
        type: integer
      created_at:
        type: string
      domain:
        type: string
      enabled:
        type: boolean
      id:
        type: integer
      remark:
        type: string
      updated_at:
        type: string
    type: object
...
  response.Any:
    properties:
      data: {}
      error_msg:
        example: ""
        type: string
    type: object
...
  tls.CertificateInput:
    properties:
      cert_pem:
        type: string
      key_pem:
        type: string
      name:
        type: string
      remark:
        type: string
    type: object
  tls.DNSAccountInput:
    properties:
      authorization:
        type: string
      name:
        type: string
      type:
        type: string
    type: object
  tls.ManagedDomainInput:
    properties:
      cert_id:
        type: integer
      domain:
        type: string
      enabled:
        type: boolean
      remark:
        type: string
    type: object
...
info:
  contact:
    name: OpenFlare
    url: https://github.com/Rain-kl/OpenFlare
  description: OpenFlare 平台后端 API，提供用户认证、系统配置、任务调度与边缘节点管理能力。
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: OpenFlare API
  version: 1.0.0
paths:
...
  /api/v1/openflare/acme-accounts/default:
    get:
      description: 返回系统默认 ACME 账号配置，需要管理员权限
      produces:
      - application/json
      responses:
        "200":
          description: 默认 ACME 账号
          schema:
            allOf:
            - $ref: '#/definitions/response.Any'
            - properties:
                data:
                  $ref: '#/definitions/model.AcmeAccount'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Any'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/response.Any'
        "403":
          description: 无管理员权限
          schema:
            $ref: '#/definitions/response.Any'
        "404":
          description: 记录不存在
          schema:
            $ref: '#/definitions/response.Any'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/response.Any'
      security:
      - SessionCookie: []
      summary: 获取默认 ACME 账号
      tags:
      - openflare-tls
...
  /api/v1/openflare/dns-accounts:
    get:
      description: 返回全部 DNS 提供商账号，需要管理员权限
      produces:
      - application/json
      responses:
        "200":
          description: DNS 账号列表
          schema:
            allOf:
            - $ref: '#/definitions/response.Any'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.DNSAccount'
                  type: array
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Any'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/response.Any'
        "403":
          description: 无管理员权限
          schema:
            $ref: '#/definitions/response.Any'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/response.Any'
      security:
      - SessionCookie: []
      summary: 列出 DNS 账号
      tags:
      - openflare-tls
    post:
      consumes:
      - application/json
      description: 创建新的 DNS 提供商账号，需要管理员权限
      parameters:
      - description: DNS 账号参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/tls.DNSAccountInput'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功的 DNS 账号
          schema:
            allOf:
            - $ref: '#/definitions/response.Any'
            - properties:
                data:
                  $ref: '#/definitions/model.DNSAccount'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Any'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/response.Any'
        "403":
          description: 无管理员权限
          schema:
            $ref: '#/definitions/response.Any'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/response.Any'
      security:
      - SessionCookie: []
      summary: 创建 DNS 账号
      tags:
      - openflare-tls
  /api/v1/openflare/dns-accounts/{id}/delete:
    post:
      description: 按 ID 删除 DNS 提供商账号，需要管理员权限
      parameters:
      - description: DNS 账号 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/response.Any'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Any'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/response.Any'
        "403":
          description: 无管理员权限
          schema:
            $ref: '#/definitions/response.Any'
        "404":
          description: 记录不存在
          schema:
            $ref: '#/definitions/response.Any'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/response.Any'
      security:
      - SessionCookie: []
      summary: 删除 DNS 账号
      tags:
      - openflare-tls
  /api/v1/openflare/dns-accounts/{id}/update:
    post:
      consumes:
      - application/json
      description: 按 ID 更新 DNS 提供商账号，需要管理员权限
      parameters:
      - description: DNS 账号 ID
        in: path
        name: id
        required: true
        type: integer
      - description: DNS 账号参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/tls.DNSAccountInput'
      produces:
      - application/json
      responses:
        "200":
          description: 更新后的 DNS 账号
          schema:
            allOf:
            - $ref: '#/definitions/response.Any'
            - properties:
                data:
                  $ref: '#/definitions/model.DNSAccount'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Any'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/response.Any'
        "403":
          description: 无管理员权限
          schema:
            $ref: '#/definitions/response.Any'
        "404":
          description: 记录不存在
          schema:
            $ref: '#/definitions/response.Any'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/response.Any'
      security:
      - SessionCookie: []
      summary: 更新 DNS 账号
      tags:
      - openflare-tls
  /api/v1/openflare/managed-domains:
    get:
      description: 返回全部托管域名及关联证书，需要管理员权限
      produces:
      - application/json
      responses:
        "200":
          description: 托管域名列表
          schema:
            allOf:
            - $ref: '#/definitions/response.Any'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.ManagedDomain'
                  type: array
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Any'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/response.Any'
        "403":
          description: 无管理员权限
          schema:
            $ref: '#/definitions/response.Any'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/response.Any'
      security:
      - SessionCookie: []
      summary: 列出托管域名
      tags:
      - openflare-tls
    post:
      consumes:
      - application/json
      description: 创建新的托管域名记录，需要管理员权限
      parameters:
      - description: 托管域名参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/tls.ManagedDomainInput'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功的托管域名
          schema:
            allOf:
            - $ref: '#/definitions/response.Any'
            - properties:
                data:
                  $ref: '#/definitions/model.ManagedDomain'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Any'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/response.Any'
        "403":
          description: 无管理员权限
          schema:
            $ref: '#/definitions/response.Any'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/response.Any'
      security:
      - SessionCookie: []
      summary: 创建托管域名
      tags:
      - openflare-tls
  /api/v1/openflare/managed-domains/{id}/delete:
    post:
      description: 按 ID 删除托管域名，需要管理员权限
      parameters:
      - description: 托管域名 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/response.Any'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Any'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/response.Any'
        "403":
          description: 无管理员权限
          schema:
            $ref: '#/definitions/response.Any'
        "404":
          description: 记录不存在
          schema:
            $ref: '#/definitions/response.Any'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/response.Any'
      security:
      - SessionCookie: []
      summary: 删除托管域名
      tags:
      - openflare-tls
  /api/v1/openflare/managed-domains/{id}/update:
    post:
      consumes:
      - application/json
      description: 按 ID 更新托管域名，需要管理员权限
      parameters:
      - description: 托管域名 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 托管域名参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/tls.ManagedDomainInput'
      produces:
      - application/json
      responses:
        "200":
          description: 更新后的托管域名
          schema:
            allOf:
            - $ref: '#/definitions/response.Any'
            - properties:
                data:
                  $ref: '#/definitions/model.ManagedDomain'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Any'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/response.Any'
        "403":
          description: 无管理员权限
          schema:
            $ref: '#/definitions/response.Any'
        "404":
          description: 记录不存在
          schema:
            $ref: '#/definitions/response.Any'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/response.Any'
      security:
      - SessionCookie: []
      summary: 更新托管域名
      tags:
      - openflare-tls
  /api/v1/openflare/managed-domains/match:
    get:
      description: 按域名查询可用的证书匹配候选，需要管理员权限
      parameters:
      - description: 域名
        in: query
        name: domain
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 证书匹配结果
          schema:
            allOf:
            - $ref: '#/definitions/response.Any'
            - properties:
                data:
                  $ref: '#/definitions/tls.ManagedDomainMatchResult'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Any'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/response.Any'
        "403":
          description: 无管理员权限
          schema:
            $ref: '#/definitions/response.Any'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/response.Any'
      security:
      - SessionCookie: []
      summary: 匹配托管域名证书
      tags:
      - openflare-tls
...
  /api/v1/openflare/tls-certificates:
    get:
      description: 返回全部 TLS 证书（不含 PEM），需要管理员权限
      produces:
      - application/json
      responses:
        "200":
          description: 证书列表
          schema:
            allOf:
            - $ref: '#/definitions/response.Any'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.TLSCertificate'
                  type: array
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Any'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/response.Any'
        "403":
          description: 无管理员权限
          schema:
            $ref: '#/definitions/response.Any'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/response.Any'
      security:
      - SessionCookie: []
      summary: 列出 TLS 证书
      tags:
      - openflare-tls
    post:
      consumes:
      - application/json
      description: 从 PEM 文本创建 TLS 证书，需要管理员权限
      parameters:
      - description: 证书参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/tls.CertificateInput'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功的证书
          schema:
            allOf:
            - $ref: '#/definitions/response.Any'
            - properties:
                data:
                  $ref: '#/definitions/model.TLSCertificate'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Any'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/response.Any'
        "403":
          description: 无管理员权限
          schema:
            $ref: '#/definitions/response.Any'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/response.Any'
      security:
      - SessionCookie: []
      summary: 创建 TLS 证书
      tags:
      - openflare-tls
  /api/v1/openflare/tls-certificates/{id}:
    get:
      description: 按 ID 返回 TLS 证书详情（不含 PEM），需要管理员权限
      parameters:
      - description: 证书 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 证书详情
          schema:
            allOf:
            - $ref: '#/definitions/response.Any'
            - properties:
                data:
                  $ref: '#/definitions/model.TLSCertificate'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Any'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/response.Any'
        "403":
          description: 无管理员权限
          schema:
            $ref: '#/definitions/response.Any'
        "404":
          description: 记录不存在
          schema:
            $ref: '#/definitions/response.Any'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/response.Any'
      security:
      - SessionCookie: []
      summary: 获取 TLS 证书详情
      tags:
      - openflare-tls
  /api/v1/openflare/tls-certificates/{id}/content:
    get:
      description: 按 ID 返回证书与私钥 PEM 内容，需要管理员权限
      parameters:
      - description: 证书 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 证书 PEM 内容
          schema:
            allOf:
            - $ref: '#/definitions/response.Any'
            - properties:
                data:
                  $ref: '#/definitions/tls.CertificateContent'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Any'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/response.Any'
        "403":
          description: 无管理员权限
          schema:
            $ref: '#/definitions/response.Any'
        "404":
          description: 记录不存在
          schema:
            $ref: '#/definitions/response.Any'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/response.Any'
      security:
      - SessionCookie: []
      summary: 获取 TLS 证书 PEM 内容
      tags:
      - openflare-tls
...
securityDefinitions:
  SessionCookie:
    in: cookie
    name: session
    type: apiKey
swagger: "2.0"